Add support for EK Certificate Chain
The EK Certificate chain can be stored in the NVRAM Indexes ranging from 0x01c00100 to 0x01c001ff, with the certificates stored concatenated in DER format.

The certificate chain will be provided by the rust-keylime agent, with the 'ek_ca_chain' key, and added to the database. Before the change the tenant checked the EK certificate against the tpm_cert_store. With the changes first the EK certificate will be checked against the ek_ca_chain and on success the EK certificate will be replaced by the most top level certificate from the chain, so the procedure can continue as before.

Signed-off-by: Eugen Matery <e.matery@digitalendoscopy.de>